requirements.txt
fastapi
uvicorn
gspread
google-auth
pandas
python-dotenv
